+++ changelog.md
@@ -1,6 +1,12 @@
-# V 2.0.2
+# Hybrid Js Utils 2
+
+## 2.0.2
 * Added tests with Jest
 * Refactor
 * README improved
 * Cleanup (removed useless files)
 * Added changelog
+
+## 2.0.3
+* Fixed countdown not setting to 0 in some cases
+* Added counter value in the return Promise for countdown